Started working on the home page
Added screens for favorite friends and viewing more information from of a profile. Think of a bounce back idea if some of this doesn't work. Calendrer and groups could become the main focus and the main screen is a list of events that are coming up and i could down size the profile creation.